+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.2.1]
+* implement getChanges and getChangesToken
+* fix datetime range passed to getRecord
+* fix getRecord implementation
 ## [1.2.0]
 * Upgrade to `alpha10` [#6](https://github.com/duynguyen242/flutter_health_connect/pull/6) by [aljkor](https://github.com/aljkor)
 ## [1.1.4]